:truck: move antispam.ts and messages.ts to utils folder
