+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "timetable-bot",
-  "version": "3.1.6",
+  "version": "3.1.7",
   "description": "",
   "main": "dist/index.js",
   "scripts": {
