+++ src/composers/mainComposer.ts
@@ -1,5 +1,5 @@
 import { Composer } from 'grammy';
-import MSGS from '../messages';
+import MSGS from '../utils/messages';
 import CustomContext from '../types/CustomContext';
 
 const composer = new Composer<CustomContext>();
